moved utils.hh/cc to utils/
