+++ bindings/bind_include.hh
@@ -29,7 +29,7 @@
 #define BINDINGS_BIND_INCLUDE_HH_
 
 #include "rascal/atomic_structure.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 #include <pybind11/eigen.h>
 #include <pybind11/iostream.h>
