+++ src/CMakeLists.txt
@@ -33,7 +33,7 @@
 set(RASCAL_SOURCES
     rascal/json_io.cc
     rascal/units.cc
-    rascal/utils.cc
+    rascal/utils/utils.cc
     rascal/utils/sparsify_fps.cc
 
     rascal/math/bessel.cc
