+++ src/rascal/representations/calculator_spherical_covariants.hh
@@ -38,7 +38,7 @@
 #include "rascal/structure_managers/property.hh"
 #include "rascal/structure_managers/property_block_sparse.hh"
 #include "rascal/structure_managers/structure_manager.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 #include <Eigen/Dense>
 #include <Eigen/Eigenvalues>
