+++ src/rascal/representations/calculator_spherical_expansion.hh
@@ -41,7 +41,7 @@
 #include "rascal/representations/cutoff_functions.hh"
 #include "rascal/structure_managers/property_block_sparse.hh"
 #include "rascal/structure_managers/structure_manager.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 #include <Eigen/Dense>
 #include <Eigen/Eigenvalues>
