+++ src/rascal/structure_managers/adaptor_center_contribution.hh
@@ -32,7 +32,7 @@
 #include "rascal/structure_managers/property.hh"
 #include "rascal/structure_managers/structure_manager.hh"
 #include "rascal/structure_managers/updateable_base.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 namespace rascal {
   /*
