+++ src/rascal/structure_managers/adaptor_half_neighbour_list.hh
@@ -32,7 +32,7 @@
 
 #include "rascal/structure_managers/property.hh"
 #include "rascal/structure_managers/structure_manager.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 namespace rascal {
   /**
