+++ src/rascal/structure_managers/adaptor_increase_maxorder.hh
@@ -34,7 +34,7 @@
 #include "rascal/lattice.hh"
 #include "rascal/structure_managers/property.hh"
 #include "rascal/structure_managers/structure_manager.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 #include <algorithm>
 #include <set>
