+++ src/rascal/structure_managers/make_structure_manager.hh
@@ -32,7 +32,7 @@
 #include "rascal/json_io.hh"
 #include "rascal/structure_managers/property.hh"
 #include "rascal/structure_managers/structure_manager.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 #include <type_traits>
