+++ src/rascal/structure_managers/property_typed.hh
@@ -35,7 +35,7 @@
 #include "rascal/math/utils.hh"
 #include "rascal/structure_managers/cluster_ref_key.hh"
 #include "rascal/structure_managers/property_base.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 namespace rascal {
