+++ src/rascal/structure_managers/structure_manager.hh
@@ -39,7 +39,7 @@
 #include "rascal/structure_managers/property.hh"
 #include "rascal/structure_managers/property_block_sparse.hh"
 #include "rascal/structure_managers/structure_manager_base.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 // Some data types and operations are based on the Eigen library
 #include <Eigen/Dense>
