+++ src/rascal/utils/utils.cc
@@ -19,7 +19,7 @@
  * Boston, MA 02111-1307, USA.
  */
 
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 #include <cstdlib>
 #include <fstream>
