+++ tests/test_utils.cc
@@ -28,7 +28,7 @@
 #include "rascal/structure_managers/cluster_ref_key.hh"
 #include "rascal/structure_managers/structure_manager.hh"
 #include "rascal/structure_managers/structure_manager_centers.hh"
-#include "rascal/utils.hh"
+#include "rascal/utils/utils.hh"
 
 #include <boost/test/unit_test.hpp>
